+++ packages/octane-is/CHANGELOG.md
@@ -1,6 +1,6 @@
-# @octanejs/react-is
+# @octanejs/octane-is
 
-## 0.1.1
+## 0.0.1
 
 ### Patch Changes